Add new Chart component
Add a new showcase for the new component
Only added basic charts configurations: no annotations, labels or tooltips
