+++ projects/showcase/src/app/components/showcase-chart/chart-types/doughnut-chart.ts
@@ -1,0 +1,17 @@
+import { ChartType } from '../../../../../../systelab-charts/src/lib/chart/interfaces';
+
+export const doughnutChart = {
+    type: ChartType.doughnut,
+    legend: {
+        enabled: false,
+    },
+    labels: ['January', 'February', 'March', 'April'],
+    datasets: [
+        {
+            data: [36, 23, 42, 52],
+            border: [{
+                width: 3,
+            }]
+        }
+    ],
+};
