+++ projects/showcase/src/app/components/showcase-chart/chart-types/polar-area-chart.ts
@@ -1,0 +1,14 @@
+import { ChartType } from '../../../../../../systelab-charts/src/lib/chart/interfaces';
+
+export const polarAreaChart = {
+    type: ChartType.polarArea,
+    datasets: [
+        {
+            label: '',
+            data: [21, 23, 42, 52],
+            border: [{
+                width: 3,
+            }]
+        },
+    ],
+};
