+++ projects/showcase/src/app/components/showcase-chart/showcase-chart.component.html
@@ -1,0 +1,55 @@
+<div class="container-fluid">
+    <div class="row mt-1">
+        <div class="col-md-6  p-2">
+            <app-showcase-title [href]="'chart'">Line Chart with added manual line</app-showcase-title>
+            <systelab-chart [config]="lineChartData"></systelab-chart>
+        </div>
+
+        <div class="col-md-6  p-2">
+            <app-showcase-title [href]="'chart'">Line Chart - Multiple Axis</app-showcase-title>
+            <systelab-chart [config]="lineMultipleAxisChartData"></systelab-chart>
+        </div>
+
+
+        <div class="col-md-6  p-2">
+            <app-showcase-title [href]="'chart'">Line Chart with timeline</app-showcase-title>
+            <systelab-chart [config]="lineChartWithTimelineData"></systelab-chart>
+            <button class="btn btn-primary btn-sm">Random TimeLine Chart Values</button>
+        </div>
+
+        <div class="col-md-6  p-2">
+            <app-showcase-title [href]="'chart'">Bar Chart</app-showcase-title>
+            <systelab-chart [config]="barChartData"></systelab-chart>
+        </div>
+
+        <div class="col-md-6 p-2">
+            <app-showcase-title [href]="'chart'">Horizontal Bar Chart</app-showcase-title>
+            <systelab-chart [config]="barChartHorizontalData"></systelab-chart>
+        </div>
+
+        <div class="col-md-6  p-2">
+            <app-showcase-title [href]="'chart'">Radar Chart</app-showcase-title>
+            <systelab-chart [config]="radarChartData"></systelab-chart>
+        </div>
+
+        <div class="col-md-6  p-2">
+            <app-showcase-title [href]="'chart'">Pie Chart</app-showcase-title>
+            <systelab-chart [config]="pieChartData"></systelab-chart>
+        </div>
+
+        <div class="col-md-6  p-2">
+            <app-showcase-title [href]="'chart'">Doughnut Chart</app-showcase-title>
+            <systelab-chart [config]="doughnutChartData"></systelab-chart>
+        </div>
+
+        <div class="col-md-6  p-2">
+            <app-showcase-title [href]="'chart'">Bubble Chart</app-showcase-title>
+            <systelab-chart [config]="bubbleChartData"></systelab-chart>
+        </div>
+
+        <div class="col-md-6  p-2">
+            <app-showcase-title [href]="'chart'">Polar Area Chart</app-showcase-title>
+            <systelab-chart [config]="polarAreaChartData"></systelab-chart>
+        </div>
+    </div>
+</div>
