+++ projects/showcase/src/app/components/showcase-charts.component.html
@@ -1,5 +1,11 @@
 <app-showcase-title [href]="'chart'">Charts</app-showcase-title>
-<app-showcase-chart></app-showcase-chart>
+
+<button (click)="selectView('legacy')">Legacy</button>
+<button (click)="selectView('new')">New</button>
+<button (click)="selectView('both')">Both</button>
+
+<app-showcase-chart *ngIf="selectedView === 'new' || selectedView === 'both'"></app-showcase-chart>
+<app-showcase-chart-legacy *ngIf="selectedView === 'legacy' || selectedView === 'both'"></app-showcase-chart-legacy>
