+++ projects/showcase/src/app/showcase.module.ts
@@ -3,10 +3,11 @@ import { NgModule } from '@angular/core';
 import { ShowcaseComponent } from './showcase.component';
 import { FormsModule } from '@angular/forms';
 import { ShowcaseTitleComponent } from './components/showcase-title.component';
-import { ShowcaseChartComponent } from './components/chart/showcase-chart.component';
+import { ShowcaseChartLegacyComponent } from './components/showcase-chart-legacy/showcase-chart-legacy.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { ShowcaseChartsComponent } from './components/showcase-charts.component';
 import { SystelabChartsModule } from 'systelab-charts';
+import { ShowcaseChartComponent } from './components/showcase-chart/showcase-chart.component';
 
 @NgModule({
 	imports:      [
@@ -18,8 +19,9 @@ import { SystelabChartsModule } from 'systelab-charts';
 	declarations: [
 		ShowcaseComponent,
 		ShowcaseTitleComponent,
-		ShowcaseChartComponent,
-		ShowcaseChartsComponent
+		ShowcaseChartLegacyComponent,
+		ShowcaseChartsComponent,
+  ShowcaseChartComponent
 	],
 	bootstrap:    [ShowcaseComponent]
 })
