+++ projects/systelab-charts/src/lib/chart-legacy/chart-legacy.component.html
@@ -1,0 +1,5 @@
+<div style="display:block;" [class.h-100]="!maintainAspectRatio && !chartResized">
+    <div #topLegend class="chart-legend"></div>
+    <canvas #canvas id="canvas" class="container-fluid"></canvas>
+    <div #bottomLegend class="chart-legend"></div>
+</div>
