+++ projects/systelab-charts/src/lib/chart-legacy/services/decimal-format.service.ts
