+++ projects/systelab-charts/src/lib/chart/chart.component.html
@@ -1,5 +1,5 @@
-<div style="display:block;" [class.h-100]="!maintainAspectRatio && !chartResized">
+<div>
     <div #topLegend class="chart-legend"></div>
-    <canvas #canvas id="canvas" class="container-fluid">{{ chart }}</canvas>
+    <canvas #chartCanvas id="chartCanvas" class="container-fluid"></canvas>
     <div #bottomLegend class="chart-legend"></div>
 </div>
