+++ projects/systelab-charts/src/lib/chart/services/legend.service.ts
@@ -1,0 +1,7 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+    providedIn: 'root',
+})
+
+export class LegendService {}
