+++ projects/systelab-charts/src/lib/chart/services/tooltip.service.ts
@@ -1,0 +1,6 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+    providedIn: 'root',
+})
+export class TooltipService {}
