+++ projects/systelab-charts/src/lib/systelab-charts.module.ts
@@ -1,7 +1,8 @@
 import { ModuleWithProviders, NgModule, Type } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ChartComponent } from './chart/chart.component';
+import { ChartLegacyComponent } from './chart-legacy/chart-legacy.component';
 import { FormsModule } from '@angular/forms';
+import { ChartComponent } from './chart/chart.component';
 
 
 
@@ -11,9 +12,12 @@ import { FormsModule } from '@angular/forms';
 		FormsModule,
 	],
 	declarations: [
+		ChartLegacyComponent,
 		ChartComponent,
+  ChartComponent,
 	],
 	exports: [
+		ChartLegacyComponent,
 		ChartComponent,
 	]
 })
