+++ projects/systelab-charts/src/public-api.ts
@@ -1,5 +1,7 @@
 
 export * from './lib/systelab-charts.module';
 export * from './lib/chart-legacy/chart-legacy.component';
-export * from './lib/services/tooltips.service';
+export * from './lib/chart/chart.component';
+export * from './lib/chart/interfaces';
+export * from './lib/chart-legacy/services/tooltip-legacy.service';
